api/requirements.txt
anyio==3.4.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
numpy==1.20.0
pandas==1.3.5
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyotp==2.6.0
python-jose==3.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.16.0
typing_extensions==4.0.1
uvicorn==0.16.0
python-multipart==0.0.5
jinja2==3.0.3
MarkupSafe==2.0.1
pytest
click
pylint
pytest-cov

